requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
anthropic
apscheduler
pyyaml
pydantic
aiosqlite
aiofiles
python-dotenv
httpx
itsdangerous
eval-type-backport
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
slack-sdk
passlib[bcrypt]
bcrypt<4.1
markdown
pytest
pytest-asyncio
